did some cleanup
