+++ Examples/Source/AppDelegate.swift
@@ -1,22 +1,12 @@
-//
-//  AppDelegate.swift
-//  iOS Example
-//
-//  Created by Leo G Dion on September 20, 2016.
-//  Copyright © 2016 BrightDigit, LLC. All rights reserved.
-//
-
 import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-  
+
   var window: UIWindow?
-  
-  func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+
+  func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
     // Override point for customization after application launch.
     return true
   }
-  
-  
 }
