+++ Examples/macOS Example/AppDelegate.swift
@@ -1,26 +1,13 @@
-//
-//  AppDelegate.swift
-//  macOS Example
-//
-//  Created by Leo Dion on 9/21/16.
-//  Copyright © 2016 BrightDigit, LLC. All rights reserved.
-//
-
 import Cocoa
 
 @NSApplicationMain
 class AppDelegate: NSObject, NSApplicationDelegate {
 
-
-
-  func applicationDidFinishLaunching(_ aNotification: Notification) {
+  func applicationDidFinishLaunching(_: Notification) {
     // Insert code here to initialize your application
   }
 
-  func applicationWillTerminate(_ aNotification: Notification) {
+  func applicationWillTerminate(_: Notification) {
     // Insert code here to tear down your application
   }
-
-
 }
-
