+++ Examples/watchOS Example Extension/RowController.swift
@@ -1,11 +1,3 @@
-//
-//  RowController.swift
-//  iOS Example
-//
-//  Created by Leo Dion on 9/21/16.
-//  Copyright © 2016 BrightDigit, LLC. All rights reserved.
-//
-
 import WatchKit
 
 class RowController: NSObject {
