+++ Tests/SwiftVerTests/SemVerTests.swift
@@ -1,27 +1,19 @@
-//
-//  SemVerTests.swift
-//  SwiftVer
-//
-//  Created by Leo Dion on 9/21/16.
-//  Copyright © 2016 BrightDigit, LLC. All rights reserved.
-//
-
 import XCTest
 @testable import SwiftVer
 
 class SemVerTests: XCTestCase {
-  
+
   override func setUp() {
     super.setUp()
     // Put setup code here. This method is called before the invocation of each test method in the class.
   }
-  
+
   override func tearDown() {
     // Put teardown code here. This method is called after the invocation of each test method in the class.
     super.tearDown()
   }
-  
-  func testGoodSemVerWithPatch () {
+
+  func testGoodSemVerWithPatch() {
     let versionString = "1.2.3"
     let semver = SemVer(versionString: versionString)
     XCTAssertEqual(semver?.major, 1)
@@ -29,8 +21,8 @@ class SemVerTests: XCTestCase {
     XCTAssertEqual(semver?.patch, 3)
     XCTAssertEqual(versionString, semver?.description)
   }
-  
-  func testGoodSemVerWithNoPatch () {
+
+  func testGoodSemVerWithNoPatch() {
     let versionString = "1.2"
     let semver = SemVer(versionString: versionString)
     XCTAssertEqual(semver?.major, 1)
@@ -38,19 +30,18 @@ class SemVerTests: XCTestCase {
     XCTAssertNil(semver?.patch)
     XCTAssertEqual(versionString, semver?.description)
   }
-  
-  func testBadSemVerString () {
+
+  func testBadSemVerString() {
     let versionString = "1.x.3"
     let semver = SemVer(versionString: versionString)
     XCTAssertNil(semver)
     XCTAssertNil(semver?.description)
   }
-  
-  func testBadSemVerValues () {
+
+  func testBadSemVerValues() {
     let versionString = "dsfasdf"
     let semver = SemVer(versionString: versionString)
     XCTAssertNil(semver)
     XCTAssertNil(semver?.description)
   }
-  
 }
