+++ Tests/SwiftVerTests/VersionControlInfoTests.swift
@@ -1,35 +1,26 @@
-//
-//  VersionControlInfoTests.swift
-//  SwiftVer
-//
-//  Created by Leo Dion on 9/21/16.
-//  Copyright © 2016 BrightDigit, LLC. All rights reserved.
-//
-
 import XCTest
 
 class VersionControlInfoTests: XCTestCase {
 
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
+  override func setUp() {
+    super.setUp()
+    // Put setup code here. This method is called before the invocation of each test method in the class.
+  }
 
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
+  override func tearDown() {
+    // Put teardown code here. This method is called after the invocation of each test method in the class.
+    super.tearDown()
+  }
 
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
+  func testExample() {
+    // This is an example of a functional test case.
+    // Use XCTAssert and related functions to verify your tests produce the correct results.
+  }
 
+  func testPerformanceExample() {
+    // This is an example of a performance test case.
+    measure {
+      // Put the code you want to measure the time of here.
+    }
+  }
 }
