+++ Tests/SwiftVerTests/VersionTests.swift
@@ -1,28 +1,20 @@
-//
-//  VersionTests.swift
-//  SwiftVer
-//
-//  Created by Leo Dion on 9/21/16.
-//  Copyright © 2016 BrightDigit, LLC. All rights reserved.
-//
-
 import XCTest
 @testable import SwiftVer
 
-struct MockBundle : VersionContainerProtocol {
-  public let infoDictionary : [String : Any]?
-  
-  public init (version: Any?, build: Any?) {
-    var infoDictionary = [String : Any]()
-    
+struct MockBundle: VersionContainerProtocol {
+  public let infoDictionary: [String: Any]?
+
+  public init(version: Any?, build: Any?) {
+    var infoDictionary = [String: Any]()
+
     if let version = version {
       infoDictionary[Version.InfoDictionaryKeys.version] = version
     }
-    
+
     if let build = build {
       infoDictionary[Version.InfoDictionaryKeys.build] = build
     }
-    
+
     self.infoDictionary = infoDictionary.count > 0 ? infoDictionary : nil
   }
 }
@@ -37,33 +29,33 @@ let versionControlInfo = VersionControlInfo(
   TAG: VCS_TAG,
   TICK: VCS_TICK,
   EXTRA: VCS_EXTRA,
-  FULL_HASH: VCS_FULL_HASH,
-  SHORT_HASH: VCS_SHORT_HASH,
-  WC_MODIFIED: VCS_WC_MODIFIED)
+  FULLHASH: VCS_FULL_HASH,
+  SHORTHASH: VCS_SHORT_HASH,
+  MODIFIED: VCS_WC_MODIFIED)
 
 class VersionTests: XCTestCase {
-  
+
   override func setUp() {
     super.setUp()
     // Put setup code here. This method is called before the invocation of each test method in the class.
   }
-  
+
   override func tearDown() {
     // Put teardown code here. This method is called after the invocation of each test method in the class.
     super.tearDown()
   }
-  
+
   func testGoodBundleVersion() {
     // This is an example of a functional test case.
     // Use XCTAssert and related functions to verify your tests produce the correct results.
     let bundle = MockBundle(version: "1.2.3", build: 4)
     let version = Version(bundle: bundle, versionControl: versionControlInfo)
-    
+
     XCTAssertEqual(version?.semver.major, 1)
     XCTAssertEqual(version?.semver.minor, 2)
     XCTAssertEqual(version?.semver.patch, 3)
     XCTAssertEqual(version?.build, 4)
-    
+
     XCTAssertEqual(version?.versionControl?.TYPE, versionControlInfo.TYPE)
     XCTAssertEqual(version?.versionControl?.BASENAME, versionControlInfo.BASENAME)
     XCTAssertEqual(version?.versionControl?.UUID, versionControlInfo.UUID)
@@ -77,45 +69,44 @@ class VersionTests: XCTestCase {
     XCTAssertEqual(version?.versionControl?.SHORT_HASH, versionControlInfo.SHORT_HASH)
     XCTAssertEqual(version?.versionControl?.WC_MODIFIED, versionControlInfo.WC_MODIFIED)
   }
-  
+
   func testMissingBundleVersion() {
     // This is an example of a functional test case.
     // Use XCTAssert and related functions to verify your tests produce the correct results.
-    
+
     let bundle = MockBundle(version: nil, build: 4)
     let version = Version(bundle: bundle, versionControl: versionControlInfo)
-    
+
     XCTAssertNil(version)
-    
   }
-  
+
   func testMissingInvalidVersion() {
     // This is an example of a functional test case.
     // Use XCTAssert and related functions to verify your tests produce the correct results.
-    
+
     let bundle = MockBundle(version: "dfasdfs", build: 4)
     let version = Version(bundle: bundle, versionControl: versionControlInfo)
-    
+
     XCTAssertNil(version)
   }
-  
+
   func testMissingBundleBuild() {
     // This is an example of a functional test case.
     // Use XCTAssert and related functions to verify your tests produce the correct results.
-    
+
     let bundle = MockBundle(version: "1.2.3", build: nil)
     let version = Version(bundle: bundle, versionControl: versionControlInfo)
-    
+
     XCTAssertNil(version)
   }
-  
+
   func testInvalidBundleBuild() {
     // This is an example of a functional test case.
     // Use XCTAssert and related functions to verify your tests produce the correct results.
-    
+
     let bundle = MockBundle(version: "1.2.3", build: "asdfdsaf")
     let version = Version(bundle: bundle, versionControl: versionControlInfo)
-    
+
     XCTAssertNil(version)
   }
 }
