+++ Tests/SwiftVerTests/testVCS.swift
@@ -1,19 +1,17 @@
-/* Generated by autorevision - do not hand-hack! */
+let VCS_TYPE = "git"
+let VCS_BASENAME = "swiftverTest"
+let VCS_UUID: String? = "1046fb20999149ca0d7a6952c0c0c4db9988087a"
+let VCS_NUM: Int = 28
+let VCS_DATE = "2016-09-21T15:12:34Z"
+let VCS_BRANCH: String = "feature/cocoapods"
+let VCS_TAG: String? = "Swift2"
+let VCS_TICK: Int? = 25
+let VCS_EXTRA: String? = nil
 
-let VCS_TYPE			= "git"
-let VCS_BASENAME		= "swiftverTest"
-let VCS_UUID:	String?	= "1046fb20999149ca0d7a6952c0c0c4db9988087a"
-let VCS_NUM:	Int		= 28
-let VCS_DATE			= "2016-09-21T15:12:34Z"
-let VCS_BRANCH:	String	= "feature/cocoapods"
-let VCS_TAG:	String?	= "Swift2"
-let VCS_TICK:	Int?	= 25
-let VCS_EXTRA:	String?	= nil
+let VCS_ACTION_STAMP: String? = "2016-09-21T15:12:34Z!leogdion@brightdigit.com"
+let VCS_FULL_HASH: String = "a22b36ca8da1e3c82a8ccb01991152a9fef6bb91"
+let VCS_SHORT_HASH: String = "a22b36c"
 
-let VCS_ACTION_STAMP:	String?	= "2016-09-21T15:12:34Z!leogdion@brightdigit.com"
-let VCS_FULL_HASH:		String	= "a22b36ca8da1e3c82a8ccb01991152a9fef6bb91"
-let VCS_SHORT_HASH:		String	= "a22b36c"
-
-let VCS_WC_MODIFIED:	Bool	= true
+let VCS_WC_MODIFIED: Bool = true
 
 /* end */
